requirements.txt
Django==3.0.2
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0